pyproject.toml
[tool.poetry]
name = "hextech-bot"
version = "0.1.0"
description = ""
authors = ["William <william1504@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
PyAutoGUI = "^0.9.53"
opencv-python = "^4.5.4"
numpy = "^1.21.4"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
